requirements.txt
PyYAML==3.12
requests==2.11.1
lxml==3.5.0
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pyOpenSSL==16.2.0